==> 1.Data_Collection.ipynb <==
# I. DATA COLLECTION 📊

Data collection inspired by [Scraping historical tweets without a Twitter Developer Account's blog](https://mihaelagrigore.medium.com/scraping-historical-tweets-without-a-twitter-developer-account-79a2c61f76ab) 

## 1. Import necessary librairies

import os
import subprocess

import json
import csv

import uuid

from IPython.display import display_javascript, display_html, display

import pandas as pd
import numpy as np

from datetime import datetime, date, time

## 2. Install snscrape and import sntwitter

pip install git+https://github.com/JustAnotherArchivist/snscrape.git

import snscrape.modules.twitter as sntwitter

## 3. Write our query

We wanted to collect tweet about Ukraine. 


Parameters from the following request :

- jsonl: This option specifies the output format as JSONL (JSON Lines). JSONL is a format where each line of the file contains a valid JSON object. It is commonly used for storing a large number of records, such as tweets, in a compact and efficient manner.

- max-results : This option sets the maximum number of results to retrieve. 

- lang : choose tweets from a specific langage

- twitter-search '#ukraine': This part of the command specifies the search query. It searches for tweets on Twitter using the specified search criteria. In this example, it searches for tweets posted with Ukraine's hashtag.

- ukraine-query-tweets.json: This redirects the output of the command to the given file. The retrieved tweets will be saved in this file in JSONL format

- since / until : to have a window which delimits tweet in time

json_filename = 'ukraine-query-tweets.json'

#Using the OS library to call CLI commands in Python
os.system(f'snscrape --max-results 200000 --jsonl --progress --since 2022-12-01 twitter-search "ukraine lang:en until:2023-06-06 " > {json_filename}')

## 4. JSON and CSV convertions

filename = 'ukraine-query-tweets'
tweets_df = pd.read_json(filename +'.json', lines=True)
tweets_df.shape

tweets_df.to_csv(filename +'.csv', index = False)

import pandas as pd

tweets_df = pd.read_csv(filename +'.csv')

display(tweets_df)

## 5. First test to see if our dataset is correct

Counting the number of Tweets we scraped


num = sum(1 for line in open(json_filename))
print(num)

Check tweets for a particular text, here "juste"


substring = 'juste'

count = 0
f = open(json_filename, 'r')
for i, line in enumerate(f):
    if substring in line:
        count = count + 1
        obj = json.loads(line)
        print(f'Tweet number {count}: {obj["content"]}')
print(count)
f.close()

tweets_df.iloc[0].content

Links mentioned in the tweet are also listed separately in the outlinks column.

tweets_df.iloc[0].outlinks


popularity_columns = ['replyCount', 'retweetCount', 'likeCount', 'quoteCount']
tweets_df.iloc[0][popularity_columns]

Find the most retweeted tweet in our dataset.

tweets_df.iloc[tweets_df.retweetCount.idxmax()][['content','retweetCount']]
