==> 4.Textual_Through_Numerical.ipynb <==
import pandas as pd

filename = 'ukraine-query-tweets-textual-rt'

# Lire le fichier CSV et créer un dataframe
tweets_df = pd.read_csv(filename +'.csv')
tweets_df = tweets_df[tweets_df['retweetCount']>20]
display(tweets_df)


from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.decomposition import PCA
import numpy as np
import pandas as pd

# Corpus de texte (colonne "rawContent") issu du DataFrame tweets_df
corpus = tweets_df["rawContent"].tolist()

# Création d'une instance de TfidfVectorizer
vectorizer = TfidfVectorizer()

# Calcul des scores TF-IDF pour le corpus
tfidf_scores = vectorizer.fit_transform(corpus)

# Réduction de la dimensionnalité avec PCA
pca = PCA(n_components=100)  # Choisir le nombre de composantes souhaité
feature_matrix_pca = pca.fit_transform(tfidf_scores.toarray())

# Création d'un nouveau DataFrame avec les features réduites
feature_df = pd.DataFrame(feature_matrix_pca, columns=[f"feature_{i+1}" for i in range(feature_matrix_pca.shape[1])])

# Ajout des index du DataFrame original au DataFrame des features
feature_df.index = tweets_df.index

# Concaténation du DataFrame original avec le DataFrame des features
tf_idf_df = pd.concat([tweets_df, feature_df], axis=1)

# Affichage du DataFrame avec les features
print(tf_idf_df)


feature_df

filename = "ukraine-query-tweets-TF-IDF"

tf_idf_df

tf_idf_df = tf_idf_df.drop('rawContent', axis=1)

tf_idf_df.to_csv('tf_idf.csv', index = False)

from gensim.models import Doc2Vec
from gensim.models.doc2vec import TaggedDocument
import numpy as np
import pandas as pd

# Corpus de texte (colonne "rawContent") issu du DataFrame tweets_df
corpus = tweets_df["rawContent"].tolist()

# Prétraitement du corpus (tokenization)
tokenized_corpus = [sentence.split() for sentence in corpus]

# Création de documents étiquetés pour l'entraînement de Doc2Vec
documents = [TaggedDocument(doc, [i]) for i, doc in enumerate(tokenized_corpus)]

# Entraînement du modèle Doc2Vec
model = Doc2Vec(documents, vector_size=100, window=5, min_count=1, workers=4)

# Création du tableau de features
feature_matrix = np.zeros((len(tokenized_corpus), model.vector_size))
for i in range(len(tokenized_corpus)):
    feature_matrix[i] = model.infer_vector(tokenized_corpus[i])

# Création d'un nouveau DataFrame avec les features
feature_df = pd.DataFrame(feature_matrix, columns=[f"feature_{i+1}" for i in range(model.vector_size)])

# Ajout des index du DataFrame original au DataFrame des features
feature_df.index = tweets_df.index

# Concaténation du DataFrame original avec le DataFrame des features
doc2vec_df = pd.concat([tweets_df, feature_df], axis=1)

# Affichage du DataFrame avec les features
print(doc2vec_df)


filename = "ukraine-query-tweets-DOC2VEC"

doc2vec_df = doc2vec_df.drop('rawContent', axis=1)

doc2vec_df.to_csv('doc2vec.csv', index = False)
