==> 5.Combinaison.ipynb <==
This file create the combinaison.csv file

from sklearn.svm import SVR
from sklearn.ensemble import RandomForestRegressor
from sklearn.neural_network import MLPRegressor
from sklearn.naive_bayes import MultinomialNB
from sklearn.metrics import mean_squared_error, r2_score

import pandas as pd

tweets_df_numerical = pd.read_csv('numerical.csv')
tweets_df_doc2vec = pd.read_csv('doc2vec.csv')
tweets_df_tfidf = pd.read_csv('tf_idf.csv')

tf_idf_doc2vec = pd.concat([tweets_df_doc2vec, tweets_df_tfidf], axis=1)

tf_idf_doc2vec.retweetCount

tf_idf_doc2vec = tf_idf_doc2vec.loc[:, ~tf_idf_doc2vec.columns.duplicated()]
tf_idf_doc2vec.retweetCount

tf_idf_doc2vec.to_csv('tf_idf_doc2vec.csv', index=False)

combinaison = pd.concat([tweets_df_numerical, tf_idf_doc2vec], axis=1)

combinaison.retweetCount

combinaison = combinaison.loc[:, ~combinaison.columns.duplicated()]
combinaison.retweetCount

combinaison.to_csv('combinaison.csv', index=False)
